requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
Django==3.1.2
django-author==1.0.2
django-bootstrap==0.2.4
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-utils-six==2.0
docopt==0.6.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.1
num2words==0.5.10
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
setuptools-git==1.2
six==1.15.0
sqlparse==0.4.1
virtualenv==20.0.33
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
